+++ lib/Event/Event.d.ts
@@ -5,11 +5,8 @@ export declare class Event {
     activeWatchers: Set<Watch | Cache>;
     activeWatcher: Watch;
     add(target: Watch | Cache): void;
-    private _start;
     start(): void;
-    private _end;
     end(): void;
-    pipe(watcher: Watch | Cache): void;
     update(): void;
 }
 export default Event;