+++ lib/Watch/Watch.d.ts
@@ -6,12 +6,12 @@ export interface Destructor<R = any> {
 }
 export declare class Watch {
     private readonly watcher;
-    static get activeWatcher(): Watch;
-    static set activeWatcher(watcher: Watch);
     destructors: Destructor[];
     private ran;
     constructor(watcher: Watcher, freeParent?: boolean, freeUpdate?: boolean);
     protected run(): any;
+    protected watchRun(): void;
+    protected forceUpdate(): void;
     update(): void;
     destroy(): void;
     onDestroy(callback: Destructor): this;