+++ lib/scope/scope.d.ts
@@ -1,0 +1,5 @@
+declare const _default: {
+    activeWatcher: any;
+    activeEvent: any;
+};
+export default _default;