+++ src/scope/scope.ts
@@ -1,0 +1,4 @@
+export default {
+  activeWatcher: undefined,
+  activeEvent: undefined,
+}